Add LinuxKPI support for attaching DRM drivers
The main difference is DRM drivers attach to a vgapci bus rather than a
pci bus. However, for now we need to override the vgapci ivars with the
grandparent's ivars in order for the rest of the LinuxKPI to work. This
is a hack, but it works fine for now.