Fix keg_large_init() for internal zones
If an internal zone has an item size that is slightly smaller than a
multiple of the page size, we won't take the slab header size into
account when allocating items. Fix that by adjusting the keg's ppera in
this case.

Also take the slab header size into account when computing the number of
pages allocated for UMA at boot time.