Use a dedicated counter for laundry thread wakeups.
We currently read the pdwakeups counter to determine how many inactive
queue scans were performed by the page daemon. This is used to determine
whether to start a background laundering. However, this was done before
the page daemon had a mechanism to wake up the laundry thread, notifying
it of a scan.

It's a bit weird to read a per-CPU counter for this purpose, and I'd
like to propose a change which makes it possible for the page daemon to
scan the inactive queue multiple times without sleeping, which would
make pdwakeups unreliable. So let's use a dedicated counter.