Fix makefs inode initialization
UFS2 does lazy initialization of inodes. makefs emulates this, but fails
to take into account the fact that we might need to initialize multiple
blocks to write a given inode. If we don't, a later initialization will
clobber that inode.

Specific inode number allocation patterns are needed to hit this case,
so it's somewhat hard to test.