Fix a memory leak introduced in r328426.
ffs_sbget() may return a superblock buffer even if it fails. Moreover,
readfunc returns a dynamically allocated buffer, when we call it on
backup superblocks, we need to free the previously allocated buffer.
Also fix what I believe is a free() of an uninitialized pointer if the
ffs_sbget() call in ffs_mountfs() fails before "fs" is set.