Use UMA_ZONE_NODUMP for the DMAR map entry zone
We may attempt to allocate from this zone while dumping to a device. I
noted that dmar_pgalloc() already uses the NODUMP flag, so it seemed
reasonable to extend this approach to the problem. If it is desirable to
retain the zone items in a dump, some more complicated approaches are
possible.