Fix the background laundering mechanism.
We used to use the number of inactive queue scans as a measure of how
many clean pages were being reclaimed by the page daemon. This was used
to decide how to launder a small number of dirty pages. With r329882, we
now scan the inactive queue more frequently, freeing a smaller number of
pages each time. Thus, the number of scans isn't as useful a metric and
we now launder too often.

Rather than counting scans, just count the total number of clean pages
freed during inactive queue scans. The laundry thread adds this value
into an accumulator which gets cleared once the background laundering
threshold is reached. This way we consider only the number of pages
freed rather than the number of scans.