Use the thread's vnet when iterating over interfaces.
For now, only permit the use of interfaces from vnet0; netdump may only
be configured from the host, and it's not clear to me that interface
names are unique across vnets. This fixes a panic when configuring
netdump with VIMAGE enabled.

Also fix a resource leak that occurs if the priv_check() in set_dumper()
or clear_dumper() fails. Note that netdump_mbuf_drain() is idempotent.