Don't bump addl_page_shortage for wired pages.
Such pages are dequeued as they're encountered during the scan, so by
the time we get to the active queue scan, they should have already been
accounted for in vmd_pagequeues[PQ_INACTIVE].pq_len. With batching this
is not quite true since the physical removal of wired pages from the
page queue is deferred until the batch queue is full, but the number of
unaccounted pages is bounded by ncpu*VM_BATCHQUEUE_SIZE and in practice
will be smaller than 1*VM_BATCHQUEUE_SIZE since the page daemon thread
will likely be running on the same CPU throughout the scan.