Eliminate the "pass" variable in the page daemon control loop.
Only scan the inactive queue when the PID controller outputs a positive
value. Move the lowmem eventhandler invocation to a separate subroutine.
Fix up some stale comments.

Now we pause() for 0.1s after an inactive queue scan failed to meet its
target. In a separate change I'd like to pause() only if the active
queue scan met its target.

As a side effect, this fixes a serious bug: after r329882, we never
reset pass to 0, so we are invoking the lowmem handlers every 10
seconds even when no memory shortage is taking place.