Have pmap_enter() reuse the PV entry when updating a mapping's PA.
As discussed in D15910, with this change pmap_enter() temporarily
invalidates the old mapping so that its PV entry can be reused.  This
sidesteps the bug reported there, where the PV entry allocation causes
the old PV entry to be reclaimed, leading to a use-after-free of the
page table page.

This addresses the same bug as r329254.