arm64: Reuse the PV entry when updating a mapping in pmap_enter().
Because arm64 updates page table entries using break-before-make, it's
not susceptible to the problem fixed by r329254.  However, it's still
susceptible to the problem with PV entry reclamation fixed in r335784
for amd64.  Fix that issue in the same way that we did for amd64.

I note that pmap_update_entry() does not return the old value of the L3
entry after invalidation, so I believe there was a small race where the
accessed or modified bits could have been set by the hardware between
the read of orig_l3 and the invalidation.  This change addresses that
too.

The original code disables interrupts during the update.  I don't see a
reason to do that for userspace mappings, so we don't do that in this
change.