Fix KVA handling in kmem_bootstrap_free().
The kmem_* layer must handle pmap updates itself, since vm_map_remove()
won't update the kernel pmap for us.  Thus, modify kmem_bootstrap_free()
to remove the KVA range from kernel_pmap and add it to kernel_arena.
For this to work, we must support demotion of the bootstrapped kernel
page tables.  It suffices to ensure that PG_A is set on the PDEs (i386
already does this) and to insert the shadowed PTPs into the kernel
pmap's radix tree when initializing them in pmap_init().