Tighten checks in the x86 stack unwinder.
With the 4/4 split on i386, INKERNEL is always true, and so we do no
validation of the first frame pointer.  This pointer will be NULL for a
thread that has been created but was never scheduled, so procstat -kka
reliably panics on i386.  Fix the problem by reordering some checks in
the unwind loop: ensure that each frame pointer, including the first,
lies within the bounds of the kernel stack.