Add some accounting to the per-domain full bucket caches.
Keep track of the number of items in each cache and use the existing UMA
callout to maintain a working set size estimate for each cache.  This is
based on section 3.7 of Bonwick's vmem paper.  The stored estimate is a
decaying average over the last 100s of activity.  The WSS
estimate is currently unused, but my aim is to use it to make
uma_reclaim() less aggressive in the common case.

The item count (uzd_nitems) has a couple of applications; it will be
useful in D7538, and could also be used to restrict the bucket cache
size of certain zones.  In particular, it would be useful for the
vm_page.c page cache zones, which are currently configured with
UMA_ZONE_NOBUCKETCACHE.