Bump pageout_oom_seq.
When an inactive queue scan fails to meet its target, we sleep
uninterruptibly.  Before r308474 we slept for a half second in between
scans, the idea being to give some time for laundering to add clean
pages to the inactive queue.  Now, we sleep for 0.1s.  The default
oom_seq value predates the new sleep period, and can result in
spurious OOM kills when swapping to slow media.  In the case that
motivated this change, a user was swapping to a USB flash drive in
addition to using it for a buildworld.  Average write latency can easily
spike to over 1s when such drives are overloaded.

The value of 100 effectively gives the kernel 10s to launder pages
before invoking the OOM killer.