Re-add the kstat.zfs.misc.arcstats.other_size sysctl.
It is used by top(1), among other things, and was removed in r337653,
where it was split into three finer-grained statistics.  We aim to
provide backward compatibility for such utilities, so re-add the
other_size sysctl under COMPAT_FREEBSD11.

The reintegration is kind of ugly, but I don't see a better way to do it
without extending the kstat KPI.