Add simple tests for SO_REUSEPORT_LB sockets.
These tests create a LB group on the loopback address, connect to the
group address a number of times, and verify that the connections were
roughly balanced among the listening sockets in the group.

This provides some basic regression testing and makes it marginally
easier to add additional SO_REUSEPORT_LB tests in the future.