Exclude the EFI framebuffer from phys_avail[].
On at least the ThunderX, the EFI memory map covers the framebuffer,
causing the kernel to treat it as regular memory.  Explicitly exclude it
during boot.

Also bump the number of excluded regions, as I'm hitting the limit.
Modify arm_physmem_exclude_region() to panic if the limit is hit instead
of silently failing.