Fix memguard when options NUMA is configured.
kmem_back() is now somewhat deficient since it doesn't know that pages
backing the same large virtual page must come from the same NUMA domain
in order to satisfy constraints of vm_reserv_extend().  It just selects
a domain according to the configured policy, which may or may not
correspond to the "colour" of the caller-supplied KVA.

memguard is the last in-tree consumer of kmem_back().  This change
allows it to work with "options NUMA" by ensuring that we consistently
select the same domain for pages backing a large virtual page.  As part
of this, ensure that the KVA range reserved for memguard is large
page-aligned.