Fix checks in the bucket allocation path.
For UMA_ZONE_NUMA zones, we only install a newly allocated bucket in the
pcpu cache if we're still running within the selected domain after the
bucket allocation.  However, if the selected domain was empty, we set
domain = UMA_ANYDOMAIN, so the test domain == PCPU_GET(domain) is always
false and we end up caching the bucket in one of the per-domain lists.

Fix the problem by testing domain == UMA_ANYDOMAIN instead of checking
for UMA_ZONE_NUMA.  As pointed out in D16666, this is clearer and saves
an extra memory reference in common cases.  After r339043, this change
also ensures that we maintain pcpu caches for CPUs belonging to extra
domains when UMA_ZONE_NUMA is set.  Presently, the only zones with this
attribute are the bucket zones themselves.