Adjust v_freecnt when blacklisting pages.
This bug has been around for a while, but the blacklisting facility was
only used if you specified pages with the vm.blacklist tunable.
However, r332182 added a new use.  This problem is particularly
pronounced on 11.2 where we expect vm_phys to return a page if
v_free_count says that there's one available.