Capsicumize savecore(8).
This is relatively straightforward.  In main(), savecore builds up a
list of device files to search for kernel dump headers.  We then use
cap_fileargs to open each device in turn.  savecore saves a number of
files in a directory (cwd by default) if a kernel dump is found; we
use a directory fd, savedirfd, to reference that directory.