Add endian.h to LHDRS.
Many applications expect to be able to include <endian.h> so this is a
compatibility sore.  Just make a symlink, like OpenBSD does.  NetBSD
uses a stub header which includes sys/endian.h, but I don't see a good
reason for that.  We use symlinks for lots of headers today.